# Improve performance of upgrade tool